requirements.txt
Flask==3.0.0
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.4
gunicorn==21.2.0
